[front/lib/api] - refactor: move getCoreViewTypeFilter to search endpoint
 - Relocated getCoreViewTypeFilter from data_source_view to the search endpoint module to streamline viewType handling
 - Removed the now-unused function from data_source_view to reduce clutter and improve module responsibilities clarity
